64-bit fixes

+++ engine/platform/win32/lib_win.h
@@ -17,7 +17,7 @@ GNU General Public License for more details.
 #include <winnt.h>
 #include <psapi.h>
 
-#define CALCULATE_ADDRESS( base, offset ) ( ( DWORD )( base ) + ( DWORD )( offset ) )
+#define CALCULATE_ADDRESS( base, offset ) ( ( size_t )( base ) + ( size_t )( offset ) )
 
 FARPROC MemoryGetProcAddress( void *module, const char *name );
 void MemoryFreeLibrary( void *hInstance );
